+++ zonemaps.h
@@ -10,7 +10,7 @@ struct zone
     std::vector<T> elements; // vector/list that holds elements in this zone
     T min;                   // minimum element in the zone
     T max;                   // maximum element in the zone
-    uint size;               // no. of elements in current zone
+    size_t size;               // no. of elements in current zone
 };
 
 template <typename T>
@@ -18,8 +18,8 @@ class zonemap
 {
     std::vector<T> elements; // vector/list that will hold all elements. Can convert to array for faster processing
     std::vector<zone<T> > zones;
-    uint num_zones;             // keeps track of the number of zones/partitions to the array/vector of elements
-    uint num_elements_per_zone; // keeps track of number of elements per zone (items in every zone has to be less than or equal to this counter)
+    size_t num_zones;             // keeps track of the number of zones/partitions to the array/vector of elements
+    size_t num_elements_per_zone; // keeps track of number of elements per zone (items in every zone has to be less than or equal to this counter)
 
 public:
     /**
@@ -27,7 +27,7 @@ public:
      * Param: number of elements per zone 
      * returns: N/A 
      */
-    zonemap(std::vector<T> _elements, uint _num_elements_per_zone);
+    zonemap(std::vector<T> _elements, size_t _num_elements_per_zone);
 
     /**
      * Purpose: builds the zonemap using the elements array/list. Maintains invariant that no. of elements in every zone
